feat(frontend): wire AutomationWorkflowStep into useWorkflowBuilder.WorkflowStep (closes #7123)
Solves both blockers from the earlier reverted attempt:

1. started_at/completed_at nullability mismatch
   Old useWorkflowBuilder.WorkflowStep used `started_at?: string` (optional)
   while AutomationWorkflowStep used `string | null` (required-but-nullable).
   Fix: AutomationWorkflowStep is now the canonical shape with
   `started_at: string | null` everywhere. The 2 object-literal call
   sites in useWorkflowBuilder.ts that were missing started_at /
   completed_at now explicitly set them to null (matches what backend
   `to_status_dict()` would emit for not-yet-started steps).

2. Duplicate WorkflowStepStatus / RiskLevel definitions
   Old: declared in both useWorkflowBuilder.ts and workflowTemplates.ts
   (two parallel string unions, same values, separate types).
   Fix: workflowTemplates.ts is the single canonical home.
   useWorkflowBuilder.ts re-exports them so existing imports
   (`useWorkflowBuilder.WorkflowStepStatus`, `useWorkflowBuilder.RiskLevel`)
   keep working unchanged.

3. AutomationWorkflowStep extended to be the runtime superset
   Added: explanation?, dependencies?, estimated_duration, execution_result?
   Tightened: status: WorkflowStepStatus (was string), risk_level required
   (was optional). Now matches what 6+ consumers expect — single source
   of truth across WorkflowRunner.vue, WorkflowLiveDashboard.vue,
   OrchestrationVisualizer.vue, WorkflowCanvas.vue, WorkflowBuilderView.vue,
   plus this composable.

4. useWorkflowBuilder.WorkflowStep collapses to alias
   `export type WorkflowStep = AutomationWorkflowStep` — all 6 consumer
   files keep their import path, no other changes needed.

Verification:

  $ npx vue-tsc --noEmit -p tsconfig.app.json
  2096 errors  # IDENTICAL to Dev_new_gui baseline; 0 new errors

The pre-existing 2096 errors are unrelated TS2307 (vue/vue-i18n module
typing issues) + implicit-any in OrchestrationVisualizer/WorkflowCanvas
arithmetic — none added by this PR, none removed.

Wire-in evidence: AutomationWorkflowStep now has 1+ production caller
(via WorkflowStep alias in useWorkflowBuilder.ts, which is consumed
by all 6 runtime view files). #6836 integration check satisfied.

Closes #7123.
